+++ CHANGELOG.md
@@ -21,6 +21,10 @@ For each release, changes are grouped under these headings:
 
 The following changes have been completed, and will be included in the next tagged release.
 
+### Fixes
+
+* Introduced `internal.ts` to resolve circular dependencies within the package.
+
 ### Dependencies
 
 * Remove dependenciy on `@ganbarodigital/ts-lib-packagename` (causes circular dependency hell)